Added register screen.

+++ src/components/Account/Register.js
@@ -10,7 +10,7 @@ class Login extends Component {
   render() {
     const POST_MUTATION = gql`
       mutation PostMutation($email: String!, $password: String!) {
-      login(input: {
+      signup(input: {
         email: $email,
         password: $password
         }) {
@@ -26,18 +26,21 @@ class Login extends Component {
 
     return (
       <Mutation mutation={POST_MUTATION}>
-        {(login) => (
+        {(signup) => (
           <div className="container pt-4">
             <div className="row justify-content-center pt-4">
               <div className="col-12 col-sm-8 col-md-6 col-lg-5">
-                <h2 className="text-center display-4">Sign-In.</h2>
+                <h2 className="text-center display-4">Register.</h2>
                 <Formik
-                  initialValues={{ email: '', password: '' }}
-                  onSubmit={async ({ email, password }, { resetForm }) => {
-                    const { data: { login: { token } } } = await login({ variables: { email, password } })
-                    localStorage.setItem("react-graphql-starter", token)
-                    resetForm()
-                    this.props.history.push("/")
+                  initialValues={{ email: '', password: '', confirmPassword: '' }}
+                  onSubmit={async ({ email, password, confirmPassword }, { resetForm }) => {
+                    debugger;
+                    if (password == confirmPassword) {
+                      const { data: { signup: { token } } } = await signup({ variables: { email, password } })
+                      localStorage.setItem("react-graphql-starter", token)
+                      resetForm()
+                      this.props.history.push("/")
+                    }
                   }}>
                   {props => {
                     const {
@@ -54,15 +57,19 @@ class Login extends Component {
                     return (
                       <form onSubmit={handleSubmit}>
                         <div className="form-label-group">
-                          <Field name="email" id="loginEmail" placeholder="Email" required value={values.email} className="form-control" type="email" />
-                          <label htmlFor="loginEmail">Email</label>
+                          <Field name="email" id="registerEmail" value={values.email} placeholder="Email" required autoFocus className="form-control" type="email" />
+                          <label htmlFor="registerEmail">Email</label>
                         </div>
                         <div className="form-label-group">
-                          <Field name="password" id="loginPassword" placeholder="Password" required value={values.password} className="form-control" type="password" />
-                          <label htmlFor="loginPassword">Password</label>
+                          <Field name="password" id="registerPassword" value={values.password} placeholder="Password" required className="form-control" type="password" />
+                          <label htmlFor="registerPassword">Password</label>
+                        </div>
+                        <div className="form-label-group">
+                          <Field name="confirmPassword" id="confirmPassword" value={values.confirmPassword} placeholder="Confirm Password" required className="form-control" type="password" />
+                          <label htmlFor="confirmPassword">Confirm Password</label>
                         </div>
                         <div className="form-group">
-                          <button className="btn btn-lg btn-primary btn-block" type="submit" disabled={!dirty || isSubmitting}>Sign-In</button>
+                          <button className="btn btn-lg btn-primary btn-block" type="submit" disabled={!dirty || isSubmitting}>Register</button>
                         </div>
                       </form>
                     );
